Use unicode in serializer backend code for better Py3k compatibility.
This affects the backend code only, not the public API. Instead of encoding
everything as it gets written, keep unicode lines and encode if necessary in
the final step. This works around Python3's restriction on concatenating
unicode and bytes instances, since everything is now unicode in its eyes. For
Python2.x, the literals are still bytes, but they get automatically converted
to unicode on concatenation so the result is the same.